VideoPress: Create VideoPress Uploader component
